Модуляция (#4)
* Добавляет Функции для тренировки

* Генерация данных для задания

* Модуляция. Разделение на модули.

---------
