+++ index.html
@@ -7,7 +7,7 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
 </head>
