+++ js/util.js
@@ -1,0 +1,23 @@
+const getRandomInteger = (a,b) => {
+  const lower = Math.ceil(Math.min(a,b));
+  const upper = Math.floor(Math.max(a,b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+const getRandomArrayElement = (items) =>
+  items[getRandomInteger(0,items.length - 1)];
+
+const createIdGenerator = () =>{
+  let numberId = 0;
+
+  return () => {
+    numberId += 1;
+    return numberId;
+  };
+};
+
+export{
+  getRandomInteger,
+  getRandomArrayElement,
+  createIdGenerator
+};
